requirements.txt
feedparser==6.0.11
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
groq>=0.9.0
